move lseek/read/write error message fns into one file
